[Uptime] Added func test for obsv location map (#61518)
* add functional test

* update func test

* refactor more changed

* update test

* update test

* update type and test

* fix the fix of fix, which din't get fix
